fix(analytics): numeric strings concatenated in concentration and dead-stock totals
customer_stats.total_revenue is NUMERIC, which node-postgres returns as a
string. The customer-concentration block summed it with +, so totalRevenue
became a 3000-char concatenated string and top5_pct/top10_pct divided by it
to 0 for every year. Customer concentration read 0% across the board.

Same class in dead-stock/search: price and ROUND(stock_qty * price) were
uncast, and capital_tied fed four separate + reduces, inflating
total_capital, dead_capital and never_sold_capital.

Retention's COUNT()/SUM() columns are coerced too — the report page offers
sorting on them and a string sort orders "1000" before "425".

+++ app/api/analytics/dead-stock/search/route.ts
@@ -73,8 +73,11 @@ export async function GET(request: Request) {
           const chainResult = (await readQueryAsync(`
             SELECT
               CAST(stock_qty AS INT) as qty,
-              price as price,
-              ROUND(stock_qty * price) as capital_tied,
+              -- ::float on both: NUMERIC comes back from node-postgres as a
+              -- string, and capital_tied feeds four `+` reduces below that
+              -- would concatenate rather than add.
+              price::float as price,
+              ROUND(stock_qty * price)::float as capital_tied,
               CAST(sold_this_year AS INT) as sold_this_year,
               CAST(sold_last_year AS INT) as sold_last_year,
               CAST(COALESCE(sold_2y_ago, 0) AS INT) as sold_2y_ago,
